finished access control feature

+++ app/Http/Controllers/AssetController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Asset;
 use App\Title;
+use App\User;
 use App\AssetStatus;
 use Str;
 use Illuminate\Http\Request;
@@ -15,8 +16,10 @@ class AssetController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Asset $asset)
     {
+        $this->authorize('viewAny',$asset);
+
         $assets = Asset::orderBy('title_id','asc')->orderBy('asset_code','asc')->get();
        return view('assets.index')->with('assets',$assets);
         //
@@ -27,8 +30,10 @@ class AssetController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create(Asset $asset)
     {
+        $this->authorize('create',$asset);
+        
         return view('assets.create')->with('titles',Title::all());
     }
 
@@ -38,14 +43,16 @@ class AssetController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(Request $request, Asset $asset)
     {
+        $this->authorize('create',$asset);
+
         $request->validate([
             'name' => 'required',
-            'qty' => 'required|min:1'
+            'quantity' => 'required|min:1'
         ]);
 
-        for($i = 0; $i < $request->input('qty'); $i++) {
+        for($i = 0; $i < $request->input('quantity'); $i++) {
            $asset = new asset;
            $asset->title_id = $request->input('name');
            $asset->asset_status_id = 1;
@@ -79,6 +86,8 @@ class AssetController extends Controller
      */
     public function show(Asset $asset)
     {
+        $this->authorize('view',$asset);
+
         return view('assets.show')->with('asset',$asset);
     }
 
@@ -91,7 +100,9 @@ class AssetController extends Controller
     public function edit(Asset $asset)
     {
         // return "Edit";
-        return view('assets.edit')->with('asset',$asset)->with('statuses',AssetStatus::all());
+        $this->authorize('update',$asset);
+
+        return view('assets.edit')->with('asset',$asset)->with('statuses',AssetStatus::all())->with('users',User::all()->whereIn('role_id',[2]));
         
     }
 
@@ -104,13 +115,21 @@ class AssetController extends Controller
      */
     public function update(Request $request, Asset $asset)
     {
+        $this->authorize('update',$asset);
+
          if ($request) {
             $request->validate([
-                'status' => 'required'
+                'status' => 'required',
+                'user' => 'required'
             ]);
         }
 
         $asset->asset_status_id = $request->input('status');
+        if($request->input('user') === 'NULL') {
+            $asset->user_id = null;
+        } else {
+            $asset->user_id = $request->input('user');
+        }
         $asset->save();
         return view('assets.show')->with('asset',$asset);
     }
@@ -123,6 +142,8 @@ class AssetController extends Controller
      */
     public function destroy(Asset $asset)
     {
+        $this->authorize('delete',$asset);
+
         $asset->delete();
         return redirect(route('assets.index'));
     }
